commit_16

+++ Doodle-jumpVS/BaseEntity.h
@@ -9,14 +9,14 @@ public:
 	BaseEntity() = default;
 	BaseEntity(const std::string& pathSprite, int x, int y) ;
 	~BaseEntity();
-	int getX() const { return m_x; }
-	int getY() const { return m_y; }
-	void setX(int _x) { m_x = _x; }
-	void setY(int _y) { m_y = _y; }
+	double getX() const { return m_x; }
+	double getY() const { return m_y; }
+	void setX(double _x) { m_x = _x; }
+	void setY(double _y) { m_y = _y; }
 	void setSize(int width, int height);
 	void draw();
 
 protected:
 	Sprite* m_entitySprite = nullptr;
-	int m_x, m_y;
+	double m_x, m_y;
 };
